refactor(providers): split video & feed providers out of app_providers.dart
Eighth batch of #4506. Moves the 16-entity video/feed cluster to
`lib/providers/video_providers.dart`:

Leaves:
- videoFilterBuilder, videoVisibilityManager
- seenVideosService, brokenVideoTracker, videoLocalStorage
- personalEventCacheService

Keystone (80 external consumers — most-consumed provider in the codebase):
- subscriptionManager
- videoEventService

Publishers:
- videoEventPublisher, viewEventPublisher
- subscribedListVideoCache
- videoSharingService
- videoMetadataUpdateService

Repositories:
- videosRepository (uses createBlockedAuthorFilter — see below)
- likesRepository, repostsRepository

The `_createBlockedAuthorFilter` helper is renamed to public
`createBlockedAuthorFilter` and stays in `app_providers.dart` because
two non-video providers (`curatedListRepository`, `peopleListsRepository`)
also call it. `videosRepository` in the new file calls it via the
back-import.

`app_providers.dart` adds `import 'video_providers.dart';` (back-import
for `analyticsService`, `hashtagService`, `curationRepository`, and other
internal consumers of `viewEventPublisherProvider`,
`videoEventServiceProvider`, `likesRepositoryProvider`) and
`export 'video_providers.dart';` (keeps 107 external consumer files
unchanged).

Mutual import with `app_providers.dart`: new file imports for
`databaseProvider`, `profileRepositoryProvider`, `funnelcakeApiClientProvider`,
`pendingActionServiceProvider`, `uploadManagerProvider`,
`blossomUploadServiceProvider`, `savedSoundsServiceProvider`,
`ogVinerCacheServiceProvider`, `curatedListsStateProvider`,
`dmRepositoryProvider`. Dart resolves library-level mutual imports lazily.

Removed via `dart fix --apply`: 20+ now-unused service/repository imports
from `app_providers.dart`. Also removed redundant imports in the new file
(moderation/preferences/relay/auth re-exported transitively via
`app_providers.dart`).

Codegen moved: 16 *Provider chunks (~951 LoC) relocated from
`app_providers.g.dart` into `video_providers.g.dart`.

Two follow-up findings (rolled into existing lists):
1. `videoEventService` factory uses 5 `service.setX()` calls + an inline
   `addVideoObserver` observer bridge for OG Viner cache learning. Already
   tracked under #3334 (split VideoEventService into focused boundaries).
2. `personalEventCacheService.initialize` and `likesRepository.initialize`
   `catchError` log at `Log.error` for expected network/IO init failures.
   Per error_handling.md matrix, should be `Log.warning`. Joins the
   init-style inconsistency follow-up.

Refs #4506, #4339 Wave 2
